feat(ios): redesign AccountView with Viseca-inspired grouped submenus
Restructure Account screen: extract Security and Preferences into
NavigationLink submenus, fix support section colors (green → neutral),
improve danger zone dark mode contrast, add iOS version to footer.
